media: move nope.media player initialization/destruction to prefetch()/release()
To avoid stalling the render thread, the player destruction is dispatched to
the background thread.

This also forces a release()/prefetch() when a live change happen. It ensures
that resources are properly re-allocated, in the right order:
- nodes are released from the root down to the leaves
- nodes are prefetched from the leaves up to the root

Fixes a crash on Android caused by incorrect resource management after media
node re-initialization caused by a live change.